Sandaruwan

+++ src/main/java/com/g7/ercauthservice/service/impl/TokenStoreServiceImpl.java
@@ -52,7 +52,9 @@ public class TokenStoreServiceImpl implements TokenStoreService {
 
     @Override
     public Token storeToken(Token token) {
-        tokenRepository.deleteTokenByUserIdAndIssueFor(token.getUserId(),token.getIssueFor());
+        if(token.getIssueFor() != IssueType.FOR_EMAIL_VERIFICATION){
+            tokenRepository.deleteTokenByUserIdAndIssueFor(token.getUserId(),token.getIssueFor());
+        }
         return tokenRepository.save(token);
     }
